requirements.txt
flask
nltk
PyPDF2
python-docx